change: renombrar NumeroGrande a EnteroGrande
