+++ packages/core/src/plugins/index.ts
@@ -15,7 +15,7 @@ import * as navegador from './navegador'
 import * as ventana from './ventana'
 import * as puente from './puente'
 import * as simbolo from './simbolo'
-import * as numerogrande from './numerogrande'
+import * as enterogrande from './enterogrande'
 import * as documento from './documento'
 
 export const plugins = [
@@ -35,7 +35,7 @@ export const plugins = [
   ['ventana', ventana],
   ['puente', puente],
   ['simbolo', simbolo],
-  ['numerogrande', numerogrande],
+  ['enterogrande', enterogrande],
   ['documento', documento]
 ]
