+++ packages/core/test/fixtures/extras/enterogrande.esjs
@@ -1,0 +1,7 @@
+mut myNum = crear EnteroGrande(1000)
+myNum.etiquetaCadena;
+myNum.aCadenaLocalizada()
+myNum.aCadena()
+myNum.valorDe()
+EnteroGrande.comoNumeroSig(64, myNum)
+EnteroGrande.comoNumeroSinN(64, myNum)
