+++ packages/core/test/fixtures/extras/enterogrande.js
@@ -1,0 +1,7 @@
+let myNum = new BigInt(1000)
+myNum.toStringTag;
+myNum.toLocaleString()
+myNum.toString()
+myNum.valueOf()
+BigInt.asIntN(64, myNum)
+BigInt.asUintN(64, myNum)
