--- packages/core/test/fixtures/extras/numerogrande.js
@@ -1,7 +1,0 @@
-let myNum = new BigInt(1000)
-myNum.toStringTag;
-myNum.toLocaleString()
-myNum.toString()
-myNum.valueOf()
-BigInt.asIntN(64, myNum)
-BigInt.asUintN(64, myNum)
